+++ src/compiler/emitter/mod.rs
@@ -4,5 +4,7 @@ mod lines_1000_1500;
 pub use lines_1000_1500::*;
 mod lines_1500_2000;
 pub use lines_1500_2000::*;
-mod lines_2000_end;
-pub use lines_2000_end::*;
+mod lines_2000_2500;
+pub use lines_2000_2500::*;
+mod lines_2500_end;
+pub use lines_2500_end::*;
